// frontend/src/core/api/generated/model/movementRequestPurpose.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

/**
 * @nullable
 */
export type MovementRequestPurpose =
  (typeof MovementRequestPurpose)[keyof typeof MovementRequestPurpose] | null

export const MovementRequestPurpose = {
  SOLD: 'SOLD',
  LOANED: 'LOANED',
  SUPPORT: 'SUPPORT',
} as const