// frontend/src/core/api/generated/model/movementRowStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

export type MovementRowStatus = (typeof MovementRowStatus)[keyof typeof MovementRowStatus]

export const MovementRowStatus = {
  PENDING_CHECK: 'PENDING_CHECK',
  IN_TRANSIT: 'IN_TRANSIT',
  DELIVERED: 'DELIVERED',
  COMPLETED: 'COMPLETED',
  AWAITING_RETURN: 'AWAITING_RETURN',
  PARTIALLY_RETURNED: 'PARTIALLY_RETURNED',
  RETURNED: 'RETURNED',
  CANCELLED: 'CANCELLED',
} as const