// frontend/src/core/api/generated/model/movementRowType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

export type MovementRowType = (typeof MovementRowType)[keyof typeof MovementRowType]

export const MovementRowType = {
  INBOUND: 'INBOUND',
  TO_SITE: 'TO_SITE',
  USED: 'USED',
  TRANSFER: 'TRANSFER',
  OUTBOUND: 'OUTBOUND',
  RETURN: 'RETURN',
  ADJUSTMENT: 'ADJUSTMENT',
} as const