// frontend/src/core/api/generated/model/stockRowStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

export type StockRowStatus = (typeof StockRowStatus)[keyof typeof StockRowStatus]

export const StockRowStatus = {
  NORMAL: 'NORMAL',
  CRITICAL: 'CRITICAL',
  OUT: 'OUT',
} as const